# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NostrMash API
  version: 0.15.0
  description: Native, admin, and Primal compatibility APIs.
servers:
  - url: /
...
paths:
...
  /api/v1/search:
    get:
      tags: [public]
      summary: Combined event/profile full-text search
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Search results with trust metadata (`trust_mode`, `trust_applied`, `result_scope`)
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/search/notes:
    get:
      tags: [public]
      summary: Notes-only full-text search with minimal product filters
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [relevant, latest], default: relevant }
        - in: query
          name: window
          schema: { type: string, enum: [24h, 7d] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, maximum: 5000, default: 0 }
      responses:
        '200':
          description: Notes search results with trust metadata (`trust_mode`, `trust_applied`, `result_scope`)
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/search/profiles:
    get:
      tags: [public]
      summary: Profiles-only full-text search with minimal product filters
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [relevant], default: relevant }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, maximum: 5000, default: 0 }
      responses:
        '200':
          description: Profiles search results with trust metadata (`trust_mode`, `trust_applied`, `result_scope`)
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/search/suggest:
    get:
      tags: [public]
      summary: Lightweight grouped suggestions for profiles and hashtags
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 20, default: 8 }
      responses:
        '200':
          description: Grouped search suggestions
        '400':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: OpaqueToken
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  request_id: { type: string }
                required: [code, message, request_id]
            required: [error]